move send mail from user controller to user service
